Move annotations into a separate package
