+++ presto-verifier/src/main/java/com/facebook/presto/verifier/annotation/ForTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.verifier.framework;
+package com.facebook.presto.verifier.annotation;
 
 import javax.inject.Qualifier;
 
@@ -26,6 +26,6 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 @Retention(RUNTIME)
 @Target({FIELD, PARAMETER, METHOD})
 @Qualifier
-public @interface ForControl
+public @interface ForTest
 {
 }
