+++ presto-verifier/src/main/java/com/facebook/presto/verifier/framework/VerificationManager.java
@@ -15,6 +15,8 @@ package com.facebook.presto.verifier.framework;
 
 import com.facebook.presto.sql.parser.ParsingException;
 import com.facebook.presto.sql.parser.SqlParser;
+import com.facebook.presto.verifier.annotation.ForControl;
+import com.facebook.presto.verifier.annotation.ForTest;
 import com.facebook.presto.verifier.event.VerifierQueryEvent;
 import com.facebook.presto.verifier.event.VerifierQueryEvent.EventStatus;
 import com.facebook.presto.verifier.source.SourceQuerySupplier;
